requirements.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-compressor==4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-daraja==1.2.0
django-heroku==0.3.1
django-phonenumber-field==6.3.0
django-sass-processor==1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
libsass==0.21.0
oauthlib==3.2.0
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
requests==2.28.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
